+++ tests/pjit_test.py
@@ -6713,7 +6713,7 @@ class ShardingInTypesTest(jtu.JaxTestCase):
     @partial(jax.shard_map, out_specs=P('x'), axis_names={'x'})
     def f1(x, i, j):
       x_a_j = x.at[:, j].get(out_sharding=jax.typeof(i).sharding)
-      return x.at[:, i].set(x_a_j)
+      return x.at[:, i].set(x_a_j, out_sharding=jax.typeof(x).sharding)
     f1(x,i,j)  # doesn't crash
 
   @config.numpy_rank_promotion('allow')
@@ -7026,19 +7026,17 @@ class ShardingInTypesTest(jtu.JaxTestCase):
     out = jax.grad(g)(embed, tok, tok_vmap)
     self.assertEqual(out.sharding, embed.sharding)
 
-  # TODO(yashkatariya): Add add and set once out_sharding=operand_sharding is
-  # removed from array methods x.at[y].set/add.
   @parameterized.named_parameters(
       (f'operand_{spec_name}_sharded_{op_name}', operand_spec, op)
       for (spec_name, operand_spec), (op_name, op) in itertools.product(
           (('xy', P('x', None, 'y')), ('x', P('x', None, None))),
-          (('mul', lambda x, ind, y: x.at[ind].mul(y)),
+          (('set', lambda x, ind, y: x.at[ind].set(y)),
+           ('add', lambda x, ind, y: x.at[ind].add(y)),
+           ('mul', lambda x, ind, y: x.at[ind].mul(y)),
            ('min', lambda x, ind, y: x.at[ind].min(y)),
            ('max', lambda x, ind, y: x.at[ind].max(y)),
            ('dynamic_update_slice_in_dim', lambda x, ind, y: (
                jax.lax.dynamic_update_slice_in_dim(x, y[None], ind, axis=0)))),
-          #  ('set', lambda x, ind, y: x.at[ind].set(y)),
-          #  ('add', lambda x, ind, y: x.at[ind].add(y)),
       )
   )
   @jtu.with_explicit_mesh((2, 2), ('x', 'y'))
@@ -8078,14 +8076,14 @@ class ShardingInTypesTest(jtu.JaxTestCase):
     @jax.jit
     def f1(x, i, j):
       x_a_j = x.at[:, j].get(out_sharding=jax.typeof(i).sharding)
-      return x.at[:, i].set(x_a_j)
+      return x.at[:, i].set(x_a_j, out_sharding=jax.typeof(x).sharding)
     f1(x,i,j)  # doesn't crash
 
     @jax.jit
     @jax.vmap
     def f2(x, i, j):
       x_j = x.at[j].get(out_sharding=jax.typeof(x).sharding)
-      return x.at[i].set(x_j)
+      return x.at[i].set(x_j, out_sharding=jax.typeof(x).sharding)
     f2(x,i,j)  # doesn't crash
 
   @jtu.with_explicit_mesh((4, 2), ('x', 'y'))
@@ -8506,16 +8504,16 @@ class ShardingInTypesTest(jtu.JaxTestCase):
     x_updated = xs.at[:, ids].add(scalar)
     self.assertEqual(x_updated.sharding, NamedSharding(mesh, P('x', None)))
 
-    @jax.jit
-    def f(x, ids, scalar):
-      out = x.at[:, ids].add(scalar)
+    @partial(jax.jit, static_argnames=('out_sharding',))
+    def f(x, ids, scalar, out_sharding = None):
+      out = x.at[:, ids].add(scalar, out_sharding=out_sharding)
       self.assertEqual(out.aval.sharding.spec, P('x', None))
       return out
 
     out = f(xs, ids, scalar)
     self.assertEqual(out.sharding, NamedSharding(mesh, P('x', None)))
 
-    out = f(xs, reshard(ids, P('x')), scalar)
+    out = f(xs, reshard(ids, P('x')), scalar, out_sharding=P('x', None))
     self.assertEqual(out.sharding, NamedSharding(mesh, P('x', None)))
 
   @config.numpy_dtype_promotion('standard')